refactor: split test submitter implementation via proxy
