+++ src/promptflow/promptflow/_sdk/operations/_flow_operations.py
@@ -130,7 +130,7 @@ class FlowOperations(TelemetryMixin):
         flow = load_flow(flow)
         flow.context.variant = variant
         from promptflow._constants import FlowLanguage
-        from promptflow._sdk._submitter.test_submitter import TestSubmitterViaProxy
+        from promptflow._sdk._submitter.test_submitter_via_proxy import TestSubmitterViaProxy
 
         if flow.dag.get(LANGUAGE_KEY, FlowLanguage.Python) == FlowLanguage.CSharp:
             with TestSubmitterViaProxy(flow=flow, flow_context=flow.context, client=self._client).init() as submitter:
